Moved the SpatialCoordinateTypes back to QL module
For some reason the ESQL QA tests could not load it from ESQL, only from QL. Until we figure out what is going wrong with the classloader, we will return these classes to their previous module.
